+++ src/pages/404.tsx
@@ -3,6 +3,7 @@ import Link from "next/link";
 import { FaGear } from "react-icons/fa6";
 
 import HoverOpacityButton from "@/components/Button/hover-opacity-button";
+import ConfettiCanvas from "@/components/confetti_overlay";
 import NavBar from "@/components/NavBar";
 
 export default function NotFound() {
@@ -120,150 +121,4 @@ export default function NotFound() {
       {totalDist.current > 100 && <ConfettiCanvas />}
     </>
   );
-}
-
-// memes
-const COLOURS = ["#f50f26", "#e310a7", "#1744e6", "#20e36e", "#e2fa2f"];
-const getRandomColor = () =>
-  COLOURS[Math.floor(Math.random() * COLOURS.length)];
-
-class Particle {
-  ctx: CanvasRenderingContext2D | null;
-  width: number;
-  height: number;
-  rotation: number;
-  x: number;
-  y: number;
-  velocity: number;
-  color: string;
-
-  constructor(
-    context: CanvasRenderingContext2D | null,
-    width: number,
-    height: number,
-    rotation: number,
-    x: number,
-    y: number,
-    velocity: number,
-    color: string
-  ) {
-    this.ctx = context;
-    this.width = width;
-    this.height = height;
-    this.rotation = rotation;
-    this.x = x;
-    this.y = y;
-    this.velocity = velocity;
-    this.color = color;
-  }
-  draw() {
-    if (!this.ctx) return;
-
-    // update position and rotation
-    this.rotation += 0.01;
-    this.y += this.velocity;
-
-    // draw
-    const rotate = (x: number, y: number) => ({
-      x:
-        (x - this.x) * Math.cos(this.rotation) -
-        (y - this.y) * Math.sin(this.rotation) +
-        this.x,
-      y:
-        (x - this.x) * Math.sin(this.rotation) +
-        (y - this.y) * Math.cos(this.rotation) +
-        this.y
-    });
-
-    // Draws rotated squares around a center point
-    this.ctx.fillStyle = this.color;
-    this.ctx.beginPath();
-    const tl = rotate(this.x - this.width / 2, this.y - this.height / 2);
-    const tr = rotate(this.x + this.width / 2, this.y - this.height / 2);
-    const br = rotate(this.x + this.width / 2, this.y + this.height / 2);
-    const bl = rotate(this.x - this.width / 2, this.y + this.height / 2);
-
-    this.ctx.moveTo(tl.x, tl.y);
-    this.ctx.lineTo(tr.x, tr.y);
-    this.ctx.lineTo(br.x, br.y);
-    this.ctx.lineTo(bl.x, bl.y);
-    this.ctx.closePath();
-    this.ctx.fill();
-  }
-}
-
-const ConfettiCanvas = () => {
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-
-  let particles: Particle[] = [];
-
-  function createParticles() {
-    if (!canvasRef || !canvasRef.current) return;
-
-    const context = canvasRef.current.getContext("2d");
-    particles = [];
-
-    let total = 100;
-    if (window.innerWidth > 1080) total = 400;
-    else if (window.innerWidth > 760) total = 300;
-    else if (window.innerWidth > 520) total = 200;
-
-    for (let i = 0; i < total; ++i) {
-      particles.push(
-        new Particle(
-          context,
-          randBetween(10, 15),
-          randBetween(4, 15),
-          randBetween(0, 360),
-          randBetween(0, window.innerWidth),
-          randBetween(-300, 0),
-          randBetween(2, 3),
-          getRandomColor()
-        )
-      );
-    }
-  }
-
-  function animateCanvas() {
-    const context = canvasRef.current?.getContext("2d");
-    if (!context || !canvasRef.current) return;
-
-    context.clearRect(0, 0, canvasRef.current.width, canvasRef.current.height);
-    particles.forEach((p) => p.draw());
-
-    requestAnimationFrame(animateCanvas);
-  }
-
-  useEffect(() => {
-    if (!canvasRef.current) return;
-
-    const handleResize = () => {
-      if (!canvasRef.current) return;
-
-      canvasRef.current.width = window.innerWidth;
-      canvasRef.current.height = window.innerHeight;
-    };
-
-    handleResize();
-    createParticles();
-    animateCanvas();
-
-    window.addEventListener("resize", handleResize);
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return (
-    <canvas
-      className="absolute top-0 left-0 z-50 pointer-events-none"
-      ref={canvasRef}
-    />
-  );
-};
-
-function randBetween(min: number, max: number): number {
-  if (min >= max) throw new Error("Min value must be less than max value");
-  return Math.random() * (max - min) + min;
 }
